Added The Way of Kings and Leadership & Self-Deception
Both are great books. You've read Mistborn? The Way of Kings is by the same author, so expect a thick, deep book steeped in world building. It's slow to get started, but still makes you want to keep reading.

Leadership & Self-Deception is by the Arbinger Institute. It's a book about being a good person. It's presented as a narrative, which makes it really fun to read, but the principles it outlines make you stop and think. Definitely the #1 recommendation I give to others.